+++ core/src/main/java/org/apache/carbondata/core/scan/result/iterator/RawResultIterator.java
@@ -28,15 +28,9 @@ import org.apache.carbondata.common.logging.LogServiceFactory;
 import org.apache.carbondata.core.constants.CarbonCommonConstants;
 import org.apache.carbondata.core.datastore.block.SegmentProperties;
 import org.apache.carbondata.core.keygenerator.KeyGenException;
-import org.apache.carbondata.core.metadata.datatype.DataType;
-import org.apache.carbondata.core.metadata.encoder.Encoding;
-import org.apache.carbondata.core.metadata.schema.table.column.CarbonDimension;
-import org.apache.carbondata.core.metadata.schema.table.column.CarbonMeasure;
-import org.apache.carbondata.core.scan.executor.util.RestructureUtil;
 import org.apache.carbondata.core.scan.result.RowBatch;
 import org.apache.carbondata.core.scan.wrappers.ByteArrayWrapper;
 import org.apache.carbondata.core.util.CarbonProperties;
-import org.apache.carbondata.core.util.DataTypeUtil;
 
 import org.apache.log4j.Logger;
 
@@ -47,9 +41,9 @@ import org.apache.log4j.Logger;
  */
 public class RawResultIterator extends CarbonIterator<Object[]> {
 
-  private final SegmentProperties sourceSegProperties;
+  protected final SegmentProperties sourceSegProperties;
 
-  private final SegmentProperties destinationSegProperties;
+  protected final SegmentProperties destinationSegProperties;
   /**
    * Iterator of the Batch raw result.
    */
@@ -64,15 +58,6 @@ public class RawResultIterator extends CarbonIterator<Object[]> {
   private Object[] currentRawRow = null;
   private boolean isBackupFilled = false;
 
-  // column reorder for no-dictionary column
-  private int noDictCount;
-  private int[] noDictMap;
-  // column drift
-  private final boolean hasColumnDrift;
-  private boolean[] isColumnDrift;
-  private int measureCount;
-  private DataType[] measureDataTypes;
-
   /**
    * LOGGER
    */
@@ -81,62 +66,18 @@ public class RawResultIterator extends CarbonIterator<Object[]> {
 
   public RawResultIterator(CarbonIterator<RowBatch> detailRawQueryResultIterator,
       SegmentProperties sourceSegProperties, SegmentProperties destinationSegProperties,
-      boolean isStreamingHandoff, boolean hasColumnDrift) {
+      boolean init) {
     this.detailRawQueryResultIterator = detailRawQueryResultIterator;
     this.sourceSegProperties = sourceSegProperties;
     this.destinationSegProperties = destinationSegProperties;
     this.executorService = Executors.newFixedThreadPool(1);
-    this.hasColumnDrift = hasColumnDrift;
-    if (!isStreamingHandoff) {
+
+    if (init) {
       init();
     }
   }
 
-  private void initForColumnDrift() {
-    List<CarbonDimension> noDictDims =
-        new ArrayList<>(destinationSegProperties.getDimensions().size());
-    for (CarbonDimension dimension : destinationSegProperties.getDimensions()) {
-      if (dimension.getNumberOfChild() == 0) {
-        if (!dimension.hasEncoding(Encoding.DICTIONARY)) {
-          noDictDims.add(dimension);
-        }
-      }
-    }
-    measureCount = destinationSegProperties.getMeasures().size();
-    noDictCount = noDictDims.size();
-    isColumnDrift = new boolean[noDictCount];
-    noDictMap = new int[noDictCount];
-    measureDataTypes = new DataType[noDictCount];
-    List<CarbonMeasure> sourceMeasures = sourceSegProperties.getMeasures();
-    int tableMeasureCount = sourceMeasures.size();
-    for (int i = 0; i < noDictCount; i++) {
-      for (int j = 0; j < tableMeasureCount; j++) {
-        if (RestructureUtil.isColumnMatches(true, noDictDims.get(i), sourceMeasures.get(j))) {
-          isColumnDrift[i] = true;
-          measureDataTypes[i] = sourceMeasures.get(j).getDataType();
-          break;
-        }
-      }
-      if (measureDataTypes[i] == null) {
-        isColumnDrift[i] = false;
-      }
-    }
-    int noDictIndex = 0;
-    // the column drift are at the end of measures
-    int measureIndex = measureCount + 1;
-    for (int i = 0; i < noDictCount; i++) {
-      if (isColumnDrift[i]) {
-        noDictMap[i] = measureIndex++;
-      } else {
-        noDictMap[i] = noDictIndex++;
-      }
-    }
-  }
-
-  private void init() {
-    if (hasColumnDrift) {
-      initForColumnDrift();
-    }
+  protected void init() {
     this.prefetchEnabled = CarbonProperties.getInstance().getProperty(
         CarbonCommonConstants.CARBON_COMPACTION_PREFETCH_ENABLE,
         CarbonCommonConstants.CARBON_COMPACTION_PREFETCH_ENABLE_DEFAULT).equalsIgnoreCase("true");
@@ -252,35 +193,12 @@ public class RawResultIterator extends CarbonIterator<Object[]> {
     return this.currentRawRow;
   }
 
-  private Object[] convertRow(Object[] rawRow) throws KeyGenException {
-    ByteArrayWrapper dimObject = (ByteArrayWrapper) rawRow[0];
-    byte[] dims = dimObject.getDictionaryKey();
+  protected Object[] convertRow(Object[] rawRow) throws KeyGenException {
+    byte[] dims = ((ByteArrayWrapper) rawRow[0]).getDictionaryKey();
     long[] keyArray = sourceSegProperties.getDimensionKeyGenerator().getKeyArray(dims);
     byte[] covertedBytes =
         destinationSegProperties.getDimensionKeyGenerator().generateKey(keyArray);
-    dimObject.setDictionaryKey(covertedBytes);
-    if (hasColumnDrift) {
-      // need move measure to dimension and return new row by current schema
-      byte[][] noDicts = dimObject.getNoDictionaryKeys();
-      byte[][] newNoDicts = new byte[noDictCount][];
-      for (int i = 0; i < noDictCount; i++) {
-        if (isColumnDrift[i]) {
-          newNoDicts[i] = DataTypeUtil
-              .getBytesDataDataTypeForNoDictionaryColumn(rawRow[noDictMap[i]], measureDataTypes[i]);
-        } else {
-          newNoDicts[i] = noDicts[noDictMap[i]];
-        }
-      }
-      ByteArrayWrapper newWrapper = new ByteArrayWrapper();
-      newWrapper.setDictionaryKey(covertedBytes);
-      newWrapper.setNoDictionaryKeys(newNoDicts);
-      newWrapper.setComplexTypesKeys(dimObject.getComplexTypesKeys());
-      newWrapper.setImplicitColumnByteArray(dimObject.getImplicitColumnByteArray());
-      Object[] finalRawRow = new Object[1 + measureCount];
-      finalRawRow[0] = newWrapper;
-      System.arraycopy(rawRow, 1, finalRawRow, 1, measureCount);
-      return finalRawRow;
-    }
+    ((ByteArrayWrapper) rawRow[0]).setDictionaryKey(covertedBytes);
     return rawRow;
   }
